Create abstract base class
