+++ gammapy/stats/__init__.py
@@ -1,6 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Statistics."""
-from .cash import *
 from .feldman_cousins import *
 from .fit_statistics import *
 from .fit_statistics_cython import *
